add synchronized subscriptions

+++ src/Services/ConferenceManagement/PaderConference.Core/Services/Synchronization/NotificationHandlers/SynchronizedObjectGarbageCollector.cs
@@ -8,31 +8,30 @@ using PaderConference.Core.Services.Synchronization.Notifications;
 
 namespace PaderConference.Core.Services.Synchronization.NotificationHandlers
 {
-    public class
-        ParticipantSubscriptionsRemovedNotificationHandler : INotificationHandler<
-            ParticipantSubscriptionsRemovedNotification>
+    public class SynchronizedObjectGarbageCollector : INotificationHandler<ParticipantSubscriptionsUpdatedNotification>
     {
         private readonly ISynchronizedObjectRepository _synchronizedObjectRepository;
         private readonly ISynchronizedObjectSubscriptionsRepository _subscriptionsRepository;
 
-        public ParticipantSubscriptionsRemovedNotificationHandler(
-            ISynchronizedObjectRepository synchronizedObjectRepository,
+        public SynchronizedObjectGarbageCollector(ISynchronizedObjectRepository synchronizedObjectRepository,
             ISynchronizedObjectSubscriptionsRepository subscriptionsRepository)
         {
             _synchronizedObjectRepository = synchronizedObjectRepository;
             _subscriptionsRepository = subscriptionsRepository;
         }
 
-        public async Task Handle(ParticipantSubscriptionsRemovedNotification notification,
+        public async Task Handle(ParticipantSubscriptionsUpdatedNotification notification,
             CancellationToken cancellationToken)
         {
-            var ((conferenceId, _), removedSubscriptions) = notification;
+            var ((conferenceId, _), removedSubscriptions, _) = notification;
+
+            if (!removedSubscriptions.Any()) return;
 
             var conferenceSubscriptions = await _subscriptionsRepository.GetOfConference(conferenceId);
             var activeSyncObjects = conferenceSubscriptions.SelectMany(x => x.Value ?? ImmutableList<string>.Empty)
                 .ToHashSet();
 
-            var inactiveSyncObjects = removedSubscriptions.Except(activeSyncObjects);
+            var inactiveSyncObjects = removedSubscriptions.Select(x => x.ToString()).Except(activeSyncObjects);
             foreach (var inactiveSubscription in inactiveSyncObjects)
             {
                 await _synchronizedObjectRepository.Remove(conferenceId, inactiveSubscription);
